+++ phpunit.xml.dist
@@ -4,7 +4,7 @@
 	xsi:noNamespaceSchemaLocation="https://schema.phpunit.de/8.5/phpunit.xsd"
 	backupGlobals="false"
 	backupStaticAttributes="false"
-	bootstrap="tests/bootstrap.php"
+	bootstrap="tests/unit/bootstrap.php"
 	colors="true"
 	convertDeprecationsToExceptions="true"
 	convertErrorsToExceptions="true"
@@ -20,7 +20,7 @@
 	>
 	<testsuites>
 		<testsuite name="duplicatepost">
-			<directory suffix="-test.php">./tests/</directory>
+			<directory suffix="-test.php">./tests/unit</directory>
 		</testsuite>
 	</testsuites>
