+++ tests/unit/ui/bulk-actions-test.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Yoast\WP\Duplicate_Post\Tests\UI;
+namespace Yoast\WP\Duplicate_Post\Tests\Unit\UI;
 
 use Brain\Monkey;
 use Mockery;
 use Yoast\WP\Duplicate_Post\Permissions_Helper;
-use Yoast\WP\Duplicate_Post\Tests\TestCase;
+use Yoast\WP\Duplicate_Post\Tests\Unit\TestCase;
 use Yoast\WP\Duplicate_Post\UI\Bulk_Actions;
 
 /**
