+++ tests/unit/ui/link-builder-test.php
@@ -1,11 +1,11 @@
 <?php
 
-namespace Yoast\WP\Duplicate_Post\Tests\UI;
+namespace Yoast\WP\Duplicate_Post\Tests\Unit\UI;
 
 use Brain\Monkey;
 use Mockery;
 use WP_Post;
-use Yoast\WP\Duplicate_Post\Tests\TestCase;
+use Yoast\WP\Duplicate_Post\Tests\Unit\TestCase;
 use Yoast\WP\Duplicate_Post\UI\Link_Builder;
 
 /**
