+++ tests/unit/ui/row-actions-test.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace Yoast\WP\Duplicate_Post\Tests\UI;
+namespace Yoast\WP\Duplicate_Post\Tests\Unit\UI;
 
 use Brain\Monkey;
 use Mockery;
 use WP_Post;
 use Yoast\WP\Duplicate_Post\Permissions_Helper;
-use Yoast\WP\Duplicate_Post\Tests\TestCase;
+use Yoast\WP\Duplicate_Post\Tests\Unit\TestCase;
 use Yoast\WP\Duplicate_Post\UI\Link_Builder;
 use Yoast\WP\Duplicate_Post\UI\Row_Actions;
